Added ButterKnife to SuperHeroesAdapter, extracted ViewHolder to a new class
