+++ app/src/main/java/com/karumi/katasuperheroes/ui/view/SuperHeroesAdapter.java
@@ -16,17 +16,15 @@
 
 package com.karumi.katasuperheroes.ui.view;
 
-import android.content.Context;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageView;
-import android.widget.TextView;
+
 import com.karumi.katasuperheroes.R;
 import com.karumi.katasuperheroes.model.SuperHero;
 import com.karumi.katasuperheroes.ui.presenter.SuperHeroesPresenter;
-import com.squareup.picasso.Picasso;
+
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -59,52 +57,5 @@ class SuperHeroesAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
 
   @Override public int getItemCount() {
     return superHeroes.size();
-  }
-
-  private static class SuperHeroViewHolder extends RecyclerView.ViewHolder {
-
-    private final SuperHeroesPresenter presenter;
-    private final ImageView photoImageView;
-    private final TextView nameTextView;
-    private final View avengersBadgeView;
-
-    public SuperHeroViewHolder(View itemView, SuperHeroesPresenter presenter) {
-      super(itemView);
-      this.presenter = presenter;
-      this.photoImageView = (ImageView) itemView.findViewById(R.id.iv_super_hero_photo);
-      this.nameTextView = (TextView) itemView.findViewById(R.id.tv_super_hero_name);
-      this.avengersBadgeView = itemView.findViewById(R.id.iv_avengers_badge);
-    }
-
-    public void render(SuperHero superHero) {
-      hookListeners(superHero);
-      renderSuperHeroPhoto(superHero.getPhoto());
-      renderSuperHeroName(superHero.getName());
-      renderAvengersBadge(superHero.isAvenger());
-    }
-
-    private void hookListeners(final SuperHero superHero) {
-      itemView.setOnClickListener(new View.OnClickListener() {
-        @Override public void onClick(View v) {
-          presenter.onSuperHeroClicked(superHero);
-        }
-      });
-    }
-
-    private void renderSuperHeroPhoto(String photo) {
-      Picasso.with(getContext()).load(photo).fit().centerCrop().into(photoImageView);
-    }
-
-    private void renderSuperHeroName(String name) {
-      nameTextView.setText(name);
-    }
-
-    private void renderAvengersBadge(boolean isAvenger) {
-      avengersBadgeView.setVisibility(isAvenger ? View.VISIBLE : View.GONE);
-    }
-
-    private Context getContext() {
-      return itemView.getContext();
-    }
   }
 }
